refactor(gwt): extract launch wizard runtime module
